t5703: simplify one-time-sed generation logic
In inconsistency(), we had two `git rev-parse` invocations in the
upstream of a pipe within a command substitution. In case this
invocation ever failed, its exit code would be swallowed up and we would
not know about it.

Pull the command substitutions out into variable assignments so that
their return codes are not lost.

Drop the pipe into tr because command substitutions should already strip
leading and trailing whitespace, including newlines.

Finally, convert the printf into an echo because it isn't necessary
anymore.

Signed-off-by: Denton Liu <liu.denton@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>